Ed and Chris got the SQL database to handle insertion, deletion, updating, and checking for a boardgame's existence in the database.  The SQL databases 'fetch()' routine was also modified to handle returning either a single row (based on a boardgame supplied) OR return all the rows in the database.
Also added a 'for each' loop in MainActivity that causes the database to process all the GameIDs that are returned from the interwebs.  One part that is missing is the ability to detect that a boardgame has been removed from the user's collection, and updating the database as appropriate.  The code may already do this but we didn't test it.